nit picky changes
