+++ cpp-linter/src/lib.rs
@@ -23,3 +23,26 @@ mod git;
 pub mod logger;
 pub mod rest_client;
 pub mod run;
+
+#[cfg(test)]
+pub(crate) mod test_common {
+    #![allow(clippy::unwrap_used)]
+
+    use std::{fs, path::PathBuf};
+
+    /// helper to avoid concurrent writes by executing (& processing the output of)
+    /// clang tools on the same test assets.
+    pub fn setup_tmp_workspace() -> tempfile::TempDir {
+        let tmp_workspace = tempfile::TempDir::with_prefix("cpp-linter-unit-tests_").unwrap();
+        let demo_path = tmp_workspace.path().join("demo");
+        fs::create_dir(&demo_path).unwrap();
+        for asset in ["demo.cpp", "demo.hpp"] {
+            fs::copy(
+                PathBuf::from("tests/demo").join(asset),
+                demo_path.join(asset),
+            )
+            .unwrap();
+        }
+        tmp_workspace
+    }
+}
